Revert "MenuHandlerの分離"
This reverts commit 4a8a9928108b066a90c91a36bd941bee1391c555.
